feat(step3): 피드백 적용 (MatchingMachine 제거)
