--- src/main/kotlin/lotto/domain/MatchingMachine.kt
@@ -1,18 +1,0 @@
-package lotto.domain
-
-object MatchingMachine {
-
-    fun match(
-        lotto: Set<LottoNumber>,
-        winningLotto: WinningLotto
-    ): PrizeResult {
-
-        val matchedCount = lotto.filter {
-            winningLotto.winningNumbers.contains(it)
-        }.count()
-
-        val isContainBonusNumber = lotto.contains(winningLotto.bonusNumber)
-
-        return PrizeResult.findByMatch(matchedCount, isContainBonusNumber)
-    }
-}
